dissipation stupid error
curr_dissipation is a ONE-CELL VALUE. That means that when it is not calculated  (ie, three out of every four iterations), it keeps the value of the last cell that it was caluclated for. Therefore iterations 2,3,4 out of the RK scheme all used the dissipation for cell i_max-2,j_max-2.

Now dissipation is stored in an arrayD3 called D, and is updated cell by cell every four iterations by update_D. that way when D persists, it will persist for every cell individually.